requirements.txt
streamlit
torch
tensorflow
transformers
langchain
langchain-community
langchain-google-genai
python-dotenv
supabase
faiss-cpu
pypdf